librbd: split image pre-removal logic to separate state machine
