+++ src/librbd/image/RemoveRequest.cc
@@ -8,14 +8,11 @@
 #include "librbd/ImageState.h"
 #include "librbd/Journal.h"
 #include "librbd/ObjectMap.h"
-#include "librbd/ExclusiveLock.h"
 #include "librbd/MirroringWatcher.h"
 #include "librbd/image/DetachChildRequest.h"
-#include "librbd/image/ListWatchersRequest.h"
-#include "librbd/journal/DisabledPolicy.h"
+#include "librbd/image/PreRemoveRequest.h"
 #include "librbd/journal/RemoveRequest.h"
 #include "librbd/mirror/DisableRequest.h"
-#include "librbd/operation/SnapshotRemoveRequest.h"
 #include "librbd/operation/TrimRequest.h"
 
 #define dout_subsys ceph_subsys_rbd
@@ -25,21 +22,6 @@
 
 namespace librbd {
 namespace image {
-
-namespace {
-
-bool auto_delete_snapshot(const SnapInfo& snap_info) {
-  auto snap_namespace_type = cls::rbd::get_snap_namespace_type(
-    snap_info.snap_namespace);
-  switch (snap_namespace_type) {
-  case cls::rbd::SNAPSHOT_NAMESPACE_TYPE_TRASH:
-    return true;
-  default:
-    return false;
-  }
-}
-
-} // anonymous namespace
 
 using librados::IoCtx;
 using util::create_context_callback;
@@ -82,7 +64,7 @@ void RemoveRequest<I>::send() {
 template<typename I>
 void RemoveRequest<I>::open_image() {
   if (m_image_ctx != nullptr) {
-    check_exclusive_lock();
+    pre_remove_image();
     return;
   }
 
@@ -122,246 +104,29 @@ void RemoveRequest<I>::handle_open_image(int r) {
   m_old_format = m_image_ctx->old_format;
   m_unknown_format = false;
 
-  check_exclusive_lock();
+  pre_remove_image();
 }
 
 template<typename I>
-void RemoveRequest<I>::check_exclusive_lock() {
-  if (m_image_ctx->operations_disabled) {
-    lderr(m_cct) << "image operations disabled due to unsupported op features"
-                 << dendl;
-    send_close_image(-EROFS);
-    return;
-  }
+void RemoveRequest<I>::pre_remove_image() {
+  ldout(m_cct, 5) << dendl;
 
-  ldout(m_cct, 20) << dendl;
-
-  if (m_image_ctx->exclusive_lock == nullptr) {
-    validate_image_removal();
-  } else {
-    acquire_exclusive_lock();
-  }
-}
-
-template<typename I>
-void RemoveRequest<I>::acquire_exclusive_lock() {
-  ldout(m_cct, 20) << dendl;
-
-  // do not attempt to open the journal when removing the image in case
-  // it's corrupt
-  if (m_image_ctx->test_features(RBD_FEATURE_JOURNALING)) {
-    RWLock::WLocker snap_locker(m_image_ctx->snap_lock);
-    m_image_ctx->set_journal_policy(new journal::DisabledPolicy());
-  }
-
-  using klass = RemoveRequest<I>;
-  if (m_force) {
-    Context *ctx = create_context_callback<
-      klass, &klass::handle_exclusive_lock_force>(this);
-    m_exclusive_lock = m_image_ctx->exclusive_lock;
-    m_exclusive_lock->shut_down(ctx);
-  } else {
-    Context *ctx = create_context_callback<
-      klass, &klass::handle_exclusive_lock>(this);
-    RWLock::WLocker owner_lock(m_image_ctx->owner_lock);
-    m_image_ctx->exclusive_lock->try_acquire_lock(ctx);
-  }
-}
-
-template<typename I>
-void RemoveRequest<I>::handle_exclusive_lock_force(int r) {
-  ldout(m_cct, 20) << "r=" << r << dendl;
-
-  delete m_exclusive_lock;
-  m_exclusive_lock = nullptr;
-
-  if (r < 0) {
-    lderr(m_cct) << "error shutting down exclusive lock: "
-                 << cpp_strerror(r) << dendl;
-    send_close_image(r);
-    return;
-  }
-
-  ceph_assert(m_image_ctx->exclusive_lock == nullptr);
-  validate_image_removal();
-}
-
-template<typename I>
-void RemoveRequest<I>::handle_exclusive_lock(int r) {
-  ldout(m_cct, 20) << "r=" << r << dendl;
-
-  if (r < 0 || !m_image_ctx->exclusive_lock->is_lock_owner()) {
-    lderr(m_cct) << "cannot obtain exclusive lock - not removing" << dendl;
-    send_close_image(-EBUSY);
-    return;
-  }
-
-  validate_image_removal();
-}
-
-template<typename I>
-void RemoveRequest<I>::validate_image_removal() {
-  ldout(m_cct, 20) << dendl;
-
-  if (!m_image_ctx->ignore_migrating &&
-      m_image_ctx->test_features(RBD_FEATURE_MIGRATING)) {
-    lderr(m_cct) << "image in migration state - not removing" << dendl;
-    send_close_image(-EBUSY);
-    return;
-  }
-
-  check_image_snaps();
-}
-
-template<typename I>
-void RemoveRequest<I>::check_image_snaps() {
-  ldout(m_cct, 20) << dendl;
-
-  m_image_ctx->snap_lock.get_read();
-  for (auto& snap_info : m_image_ctx->snap_info) {
-    if (auto_delete_snapshot(snap_info.second)) {
-      m_snap_infos.insert(snap_info);
-    } else {
-      m_image_ctx->snap_lock.put_read();
-
-      lderr(m_cct) << "image has snapshots - not removing" << dendl;
-      send_close_image(-ENOTEMPTY);
-      return;
-    }
-  }
-  m_image_ctx->snap_lock.put_read();
-
-  list_image_watchers();
-}
-
-template<typename I>
-void RemoveRequest<I>::list_image_watchers() {
-  ldout(m_cct, 20) << dendl;
-
-  int flags = LIST_WATCHERS_FILTER_OUT_MY_INSTANCE |
-              LIST_WATCHERS_FILTER_OUT_MIRROR_INSTANCES;
-  using klass = RemoveRequest<I>;
-  Context *ctx = create_context_callback<
-    klass, &klass::handle_list_image_watchers>(this);
-
-  auto req = ListWatchersRequest<I>::create(*m_image_ctx, flags, &m_watchers,
-                                            ctx);
-  req->send();
-}
-
-template<typename I>
-void RemoveRequest<I>::handle_list_image_watchers(int r) {
-  ldout(m_cct, 20) << "r=" << r << dendl;
-
-  if (r < 0) {
-    lderr(m_cct) << "error listing image watchers: " << cpp_strerror(r)
-                 << dendl;
-    send_close_image(r);
-    return;
-  }
-
-  check_image_watchers();
-}
-
-template<typename I>
-void RemoveRequest<I>::check_image_watchers() {
-  if (!m_watchers.empty()) {
-    lderr(m_cct) << "image has watchers - not removing" << dendl;
-    send_close_image(-EBUSY);
-    return;
-  }
-
-  check_group();
-}
-
-template<typename I>
-void RemoveRequest<I>::check_group() {
-  if (m_old_format) {
-    trim_image();
-    return;
-  }
-
-  ldout(m_cct, 20) << dendl;
-
-  librados::ObjectReadOperation op;
-  librbd::cls_client::image_group_get_start(&op);
-
-  using klass = RemoveRequest<I>;
-  librados::AioCompletion *rados_completion = create_rados_callback<
-    klass, &klass::handle_check_group>(this);
-  m_out_bl.clear();
-  int r = m_image_ctx->md_ctx.aio_operate(m_header_oid, rados_completion, &op,
-                                          &m_out_bl);
-  ceph_assert(r == 0);
-  rados_completion->release();
-}
-
-template<typename I>
-void RemoveRequest<I>::handle_check_group(int r) {
-  ldout(m_cct, 20) << "r=" << r << dendl;
-
-  cls::rbd::GroupSpec s;
-  if (r == 0) {
-    auto it = m_out_bl.cbegin();
-    r = librbd::cls_client::image_group_get_finish(&it, &s);
-  }
-  if (r < 0 && r != -EOPNOTSUPP) {
-    lderr(m_cct) << "error fetching group for image: "
-                 << cpp_strerror(r) << dendl;
-    send_close_image(r);
-    return;
-  }
-
-  if (s.is_valid()) {
-    lderr(m_cct) << "image is in a group - not removing" << dendl;
-    send_close_image(-EMLINK);
-    return;
-  }
-
-  remove_snapshot();
-}
-
-template<typename I>
-void RemoveRequest<I>::remove_snapshot() {
-  if (m_snap_infos.empty()) {
-    trim_image();
-    return;
-  }
-
-  auto snap_id = m_snap_infos.begin()->first;
-  auto& snap_info = m_snap_infos.begin()->second;
-  ldout(m_cct, 20) << "snap_id=" << snap_id << ", "
-                   << "snap_name=" << snap_info.name << dendl;
-
-  RWLock::RLocker owner_lock(m_image_ctx->owner_lock);
   auto ctx = create_context_callback<
-    RemoveRequest<I>, &RemoveRequest<I>::handle_remove_snapshot>(this);
-  auto req = librbd::operation::SnapshotRemoveRequest<I>::create(
-    *m_image_ctx, snap_info.snap_namespace, snap_info.name,
-    snap_id, ctx);
+    RemoveRequest<I>, &RemoveRequest<I>::handle_pre_remove_image>(this);
+  auto req = PreRemoveRequest<I>::create(m_image_ctx, m_force, ctx);
   req->send();
 }
 
 template<typename I>
-void RemoveRequest<I>::handle_remove_snapshot(int r) {
-  ldout(m_cct, 20) << "r=" << r << dendl;
+void RemoveRequest<I>::handle_pre_remove_image(int r) {
+  ldout(m_cct, 5) << "r=" << r << dendl;
 
-  if (r < 0 && r != -ENOENT) {
-    auto snap_id = m_snap_infos.begin()->first;
-    lderr(m_cct) << "failed to auto-prune snapshot " << snap_id << ": "
-                 << cpp_strerror(r) << dendl;
-
-    if (r == -EBUSY) {
-      r = -ENOTEMPTY;
-    }
+  if (r < 0) {
     send_close_image(r);
     return;
   }
 
-  ceph_assert(!m_snap_infos.empty());
-  m_snap_infos.erase(m_snap_infos.begin());
-
-  remove_snapshot();
+  trim_image();
 }
 
 template<typename I>
