+++ src/test/librbd/image/test_mock_RemoveRequest.cc
@@ -9,15 +9,12 @@
 #include "test/librados_test_stub/MockTestMemRadosClient.h"
 #include "librbd/ImageState.h"
 #include "librbd/internal.h"
-#include "librbd/Operations.h"
 #include "librbd/image/TypeTraits.h"
 #include "librbd/image/DetachChildRequest.h"
-#include "librbd/image/ListWatchersRequest.h"
+#include "librbd/image/PreRemoveRequest.h"
 #include "librbd/image/RemoveRequest.h"
-#include "librbd/image/RefreshParentRequest.h"
 #include "librbd/journal/RemoveRequest.h"
 #include "librbd/mirror/DisableRequest.h"
-#include "librbd/operation/SnapshotRemoveRequest.h"
 #include "librbd/operation/TrimRequest.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
@@ -46,6 +43,7 @@ struct MockTestImageCtx : public MockImageCtx {
 MockTestImageCtx* MockTestImageCtx::s_instance = nullptr;
 
 } // anonymous namespace
+
 namespace image {
 
 template <>
@@ -75,18 +73,12 @@ public:
 
 DetachChildRequest<MockTestImageCtx> *DetachChildRequest<MockTestImageCtx>::s_instance;
 
-} // namespace image
-
-namespace operation {
-
 template <>
-class SnapshotRemoveRequest<MockTestImageCtx> {
+class PreRemoveRequest<MockTestImageCtx> {
 public:
-  static SnapshotRemoveRequest *s_instance;
-  static SnapshotRemoveRequest *create(MockTestImageCtx &image_ctx,
-                                       cls::rbd::SnapshotNamespace sn,
-                                       std::string name,
-                                       uint64_t id, Context *on_finish) {
+  static PreRemoveRequest *s_instance;
+  static PreRemoveRequest *create(MockTestImageCtx* image_ctx, bool force,
+                                  Context* on_finish) {
     ceph_assert(s_instance != nullptr);
     s_instance->on_finish = on_finish;
     return s_instance;
@@ -94,14 +86,18 @@ public:
 
   Context *on_finish = nullptr;
 
-  SnapshotRemoveRequest() {
+  PreRemoveRequest() {
     s_instance = this;
   }
 
   MOCK_METHOD0(send, void());
 };
 
-SnapshotRemoveRequest<MockTestImageCtx> *SnapshotRemoveRequest<MockTestImageCtx>::s_instance;
+PreRemoveRequest<MockTestImageCtx> *PreRemoveRequest<MockTestImageCtx>::s_instance = nullptr;
+
+} // namespace image
+
+namespace operation {
 
 template <>
 class TrimRequest<MockTestImageCtx> {
@@ -183,48 +179,10 @@ public:
 DisableRequest<MockTestImageCtx> *DisableRequest<MockTestImageCtx>::s_instance;
 
 } // namespace mirror
-
-namespace image {
-
-template<>
-class ListWatchersRequest<MockTestImageCtx> {
-public:
-  static ListWatchersRequest *s_instance;
-  Context *on_finish = nullptr;
-
-  static ListWatchersRequest *create(MockTestImageCtx &image_ctx, int flags,
-                                     std::list<obj_watch_t> *watchers,
-                                     Context *on_finish) {
-    ceph_assert(s_instance != nullptr);
-    s_instance->on_finish = on_finish;
-    return s_instance;
-  }
-
-  ListWatchersRequest() {
-    s_instance = this;
-  }
-
-  MOCK_METHOD0(send, void());
-};
-
-ListWatchersRequest<MockTestImageCtx> *ListWatchersRequest<MockTestImageCtx>::s_instance;
-
-} // namespace image
 } // namespace librbd
 
 // template definitions
 #include "librbd/image/RemoveRequest.cc"
-
-ACTION_P(TestFeatures, image_ctx) {
-  return ((image_ctx->features & arg0) != 0);
-}
-
-ACTION_P(ShutDownExclusiveLock, image_ctx) {
-  // shutting down exclusive lock will close object map and journal
-  image_ctx->exclusive_lock = nullptr;
-  image_ctx->object_map = nullptr;
-  image_ctx->journal = nullptr;
-}
 
 namespace librbd {
 namespace image {
@@ -244,16 +202,14 @@ public:
   typedef ::librbd::image::TypeTraits<MockTestImageCtx> TypeTraits;
   typedef typename TypeTraits::ContextWQ ContextWQ;
   typedef RemoveRequest<MockTestImageCtx> MockRemoveRequest;
+  typedef PreRemoveRequest<MockTestImageCtx> MockPreRemoveRequest;
   typedef DetachChildRequest<MockTestImageCtx> MockDetachChildRequest;
-  typedef ListWatchersRequest<MockTestImageCtx> MockListWatchersRequest;
-  typedef librbd::operation::SnapshotRemoveRequest<MockTestImageCtx> MockSnapshotRemoveRequest;
   typedef librbd::operation::TrimRequest<MockTestImageCtx> MockTrimRequest;
   typedef librbd::journal::RemoveRequest<MockTestImageCtx> MockJournalRemoveRequest;
   typedef librbd::mirror::DisableRequest<MockTestImageCtx> MockMirrorDisableRequest;
 
   librbd::ImageCtx *m_test_imctx = NULL;
   MockTestImageCtx *m_mock_imctx = NULL;
-
 
   void SetUp() override {
     TestMockFixture::SetUp();
@@ -293,29 +249,10 @@ public:
                 }));
   }
 
-  void expect_list_image_watchers(
-    MockTestImageCtx &mock_image_ctx,
-    MockListWatchersRequest &mock_list_watchers_request, int r) {
-    EXPECT_CALL(mock_list_watchers_request, send())
-      .WillOnce(FinishRequest(&mock_list_watchers_request, r, &mock_image_ctx));
-  }
-
-  void expect_get_group(MockTestImageCtx &mock_image_ctx, int r) {
-    auto &expect = EXPECT_CALL(get_mock_io_ctx(mock_image_ctx.md_ctx),
-                               exec(mock_image_ctx.header_oid, _, StrEq("rbd"),
-                                    StrEq("image_group_get"), _, _, _));
-    if (r < 0) {
-      expect.WillOnce(Return(r));
-    } else {
-      expect.WillOnce(DoDefault());
-    }
-  }
-
-  void expect_remove_snap(MockTestImageCtx &mock_image_ctx,
-                          MockSnapshotRemoveRequest& mock_snap_remove_request,
-                          int r) {
-    EXPECT_CALL(mock_snap_remove_request, send())
-      .WillOnce(FinishRequest(&mock_snap_remove_request, r, &mock_image_ctx));
+  void expect_pre_remove_image(MockTestImageCtx &mock_image_ctx,
+                               MockPreRemoveRequest& mock_request, int r) {
+    EXPECT_CALL(mock_request, send())
+      .WillOnce(FinishRequest(&mock_request, r, &mock_image_ctx));
   }
 
   void expect_trim(MockTestImageCtx &mock_image_ctx,
@@ -355,32 +292,6 @@ public:
     EXPECT_CALL(mock_request, send())
       .WillOnce(FinishRequest(&mock_request, r, &mock_image_ctx));
   }
-
-  void expect_test_features(MockTestImageCtx &mock_image_ctx) {
-    EXPECT_CALL(mock_image_ctx, test_features(_))
-      .WillRepeatedly(TestFeatures(&mock_image_ctx));
-  }
-
-  void expect_set_journal_policy(MockTestImageCtx &mock_image_ctx) {
-    if (m_test_imctx->test_features(RBD_FEATURE_JOURNALING)) {
-      EXPECT_CALL(mock_image_ctx, set_journal_policy(_))
-        .WillOnce(Invoke([](journal::Policy* policy) {
-                    ASSERT_TRUE(policy->journal_disabled());
-                    delete policy;
-                  }));
-    }
-  }
-
-  void expect_shut_down_exclusive_lock(MockTestImageCtx &mock_image_ctx,
-                                       MockExclusiveLock &mock_exclusive_lock,
-                                       int r) {
-    if (m_mock_imctx->exclusive_lock != nullptr) {
-      EXPECT_CALL(mock_exclusive_lock, shut_down(_))
-        .WillOnce(DoAll(ShutDownExclusiveLock(&mock_image_ctx),
-                        CompleteContext(r, mock_image_ctx.image_ctx->op_work_queue)));
-    }
-  }
-
 };
 
 TEST_F(TestMockImageRemoveRequest, SuccessV1) {
@@ -389,10 +300,9 @@ TEST_F(TestMockImageRemoveRequest, SuccessV1) {
 
   InSequence seq;
   expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
 
-  MockListWatchersRequest mock_list_watchers_request;
-  expect_list_image_watchers(*m_mock_imctx, mock_list_watchers_request, 0);
+  MockPreRemoveRequest mock_pre_remove_request;
+  expect_pre_remove_image(*m_mock_imctx, mock_pre_remove_request, 0);
 
   MockTrimRequest mock_trim_request;
   expect_trim(*m_mock_imctx, mock_trim_request, 0);
@@ -432,11 +342,6 @@ TEST_F(TestMockImageRemoveRequest, OpenFailV1) {
 TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV1) {
   REQUIRE_FEATURE(RBD_FEATURE_LAYERING);
 
-  MockExclusiveLock *mock_exclusive_lock = new MockExclusiveLock();
-  if (m_test_imctx->test_features(RBD_FEATURE_EXCLUSIVE_LOCK)) {
-    m_mock_imctx->exclusive_lock = mock_exclusive_lock;
-  }
-
   expect_op_work_queue(*m_mock_imctx);
 
   m_mock_imctx->parent_md.spec.pool_id = m_ioctx.get_id();
@@ -445,20 +350,9 @@ TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV1) {
 
   InSequence seq;
   expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
 
-  if (m_mock_imctx->exclusive_lock != nullptr) {
-    expect_test_features(*m_mock_imctx);
-  }
-  expect_set_journal_policy(*m_mock_imctx);
-  expect_shut_down_exclusive_lock(*m_mock_imctx, *mock_exclusive_lock, 0);
-
-  expect_test_features(*m_mock_imctx);
-
-  MockListWatchersRequest mock_list_watchers_request;
-  expect_list_image_watchers(*m_mock_imctx, mock_list_watchers_request, 0);
-
-  expect_get_group(*m_mock_imctx, 0);
+  MockPreRemoveRequest mock_pre_remove_request;
+  expect_pre_remove_image(*m_mock_imctx, mock_pre_remove_request, 0);
 
   MockTrimRequest mock_trim_request;
   expect_trim(*m_mock_imctx, mock_trim_request, 0);
@@ -490,11 +384,6 @@ TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV1) {
 TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV2) {
   REQUIRE_FEATURE(RBD_FEATURE_LAYERING);
 
-  MockExclusiveLock *mock_exclusive_lock = new MockExclusiveLock();
-  if (m_test_imctx->test_features(RBD_FEATURE_EXCLUSIVE_LOCK)) {
-    m_mock_imctx->exclusive_lock = mock_exclusive_lock;
-  }
-
   expect_op_work_queue(*m_mock_imctx);
 
   m_mock_imctx->parent_md.spec.pool_id = m_ioctx.get_id();
@@ -503,20 +392,9 @@ TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV2) {
 
   InSequence seq;
   expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
 
-  if (m_mock_imctx->exclusive_lock != nullptr) {
-    expect_test_features(*m_mock_imctx);
-  }
-  expect_set_journal_policy(*m_mock_imctx);
-  expect_shut_down_exclusive_lock(*m_mock_imctx, *mock_exclusive_lock, 0);
-
-  expect_test_features(*m_mock_imctx);
-
-  MockListWatchersRequest mock_list_watchers_request;
-  expect_list_image_watchers(*m_mock_imctx, mock_list_watchers_request, 0);
-
-  expect_get_group(*m_mock_imctx, 0);
+  MockPreRemoveRequest mock_pre_remove_request;
+  expect_pre_remove_image(*m_mock_imctx, mock_pre_remove_request, 0);
 
   MockTrimRequest mock_trim_request;
   expect_trim(*m_mock_imctx, mock_trim_request, 0);
@@ -548,11 +426,6 @@ TEST_F(TestMockImageRemoveRequest, SuccessV2CloneV2) {
 TEST_F(TestMockImageRemoveRequest, NotExistsV2) {
   REQUIRE_FEATURE(RBD_FEATURE_JOURNALING);
 
-  MockExclusiveLock *mock_exclusive_lock = new MockExclusiveLock();
-  if (m_test_imctx->test_features(RBD_FEATURE_EXCLUSIVE_LOCK)) {
-    m_mock_imctx->exclusive_lock = mock_exclusive_lock;
-  }
-
   expect_op_work_queue(*m_mock_imctx);
 
   m_mock_imctx->parent_md.spec.pool_id = m_ioctx.get_id();
@@ -561,18 +434,9 @@ TEST_F(TestMockImageRemoveRequest, NotExistsV2) {
 
   InSequence seq;
   expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
 
-  expect_test_features(*m_mock_imctx);
-  expect_set_journal_policy(*m_mock_imctx);
-  expect_shut_down_exclusive_lock(*m_mock_imctx, *mock_exclusive_lock, 0);
-
-  expect_test_features(*m_mock_imctx);
-
-  MockListWatchersRequest mock_list_watchers_request;
-  expect_list_image_watchers(*m_mock_imctx, mock_list_watchers_request, 0);
-
-  expect_get_group(*m_mock_imctx, 0);
+  MockPreRemoveRequest mock_pre_remove_request;
+  expect_pre_remove_image(*m_mock_imctx, mock_pre_remove_request, 0);
 
   MockTrimRequest mock_trim_request;
   expect_trim(*m_mock_imctx, mock_trim_request, 0);
@@ -598,127 +462,6 @@ TEST_F(TestMockImageRemoveRequest, NotExistsV2) {
     m_ioctx, m_image_name, "", true, false, no_op, &op_work_queue, &ctx);
   req->send();
   ASSERT_EQ(-ENOENT, ctx.wait());
-}
-
-TEST_F(TestMockImageRemoveRequest, OperationsDisabled) {
-  MockExclusiveLock mock_exclusive_lock;
-  if (m_test_imctx->test_features(RBD_FEATURE_EXCLUSIVE_LOCK)) {
-    m_mock_imctx->exclusive_lock = &mock_exclusive_lock;
-  }
-
-  m_mock_imctx->operations_disabled = true;
-
-  InSequence seq;
-  expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
-  expect_state_close(*m_mock_imctx);
-
-  C_SaferCond ctx;
-  librbd::NoOpProgressContext no_op;
-  ContextWQ op_work_queue;
-  MockRemoveRequest *req = MockRemoveRequest::create(
-    m_ioctx, m_image_name, "", true, false, no_op, &op_work_queue, &ctx);
-  req->send();
-
-  ASSERT_EQ(-EROFS, ctx.wait());
-}
-
-TEST_F(TestMockImageRemoveRequest, Migration) {
-  m_mock_imctx->features |= RBD_FEATURE_MIGRATING;
-
-  InSequence seq;
-  expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
-  expect_state_close(*m_mock_imctx);
-
-  C_SaferCond ctx;
-  librbd::NoOpProgressContext no_op;
-  ContextWQ op_work_queue;
-  MockRemoveRequest *req = MockRemoveRequest::create(
-    m_ioctx, m_image_name, "", true, false, no_op, &op_work_queue, &ctx);
-  req->send();
-
-  ASSERT_EQ(-EBUSY, ctx.wait());
-}
-
-TEST_F(TestMockImageRemoveRequest, Snapshots) {
-  m_mock_imctx->snap_info = {
-    {123, {"snap1", {cls::rbd::UserSnapshotNamespace{}}, {}, {}, {}, {}, {}}}};
-
-  InSequence seq;
-  expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
-  expect_state_close(*m_mock_imctx);
-
-  C_SaferCond ctx;
-  librbd::NoOpProgressContext no_op;
-  ContextWQ op_work_queue;
-  MockRemoveRequest *req = MockRemoveRequest::create(
-    m_ioctx, m_image_name, "", true, false, no_op, &op_work_queue, &ctx);
-  req->send();
-
-  ASSERT_EQ(-ENOTEMPTY, ctx.wait());
-}
-
-TEST_F(TestMockImageRemoveRequest, AutoDeleteSnapshots) {
-  REQUIRE_FORMAT_V2();
-
-  MockExclusiveLock *mock_exclusive_lock = new MockExclusiveLock();
-  if (m_test_imctx->test_features(RBD_FEATURE_EXCLUSIVE_LOCK)) {
-    m_mock_imctx->exclusive_lock = mock_exclusive_lock;
-  }
-
-  expect_op_work_queue(*m_mock_imctx);
-
-  m_mock_imctx->snap_info = {
-    {123, {"snap1", {cls::rbd::TrashSnapshotNamespace{}}, {}, {}, {}, {}, {}}}};
-
-  InSequence seq;
-  expect_state_open(*m_mock_imctx, 0);
-  expect_test_features(*m_mock_imctx);
-
-  if (m_mock_imctx->exclusive_lock != nullptr) {
-    expect_test_features(*m_mock_imctx);
-  }
-  expect_set_journal_policy(*m_mock_imctx);
-  expect_shut_down_exclusive_lock(*m_mock_imctx, *mock_exclusive_lock, 0);
-
-  expect_test_features(*m_mock_imctx);
-
-  MockListWatchersRequest mock_list_watchers_request;
-  expect_list_image_watchers(*m_mock_imctx, mock_list_watchers_request, 0);
-
-  expect_get_group(*m_mock_imctx, 0);
-
-  MockSnapshotRemoveRequest mock_snap_remove_request;
-  expect_remove_snap(*m_mock_imctx, mock_snap_remove_request, 0);
-
-  MockTrimRequest mock_trim_request;
-  expect_trim(*m_mock_imctx, mock_trim_request, 0);
-
-  MockDetachChildRequest mock_detach_child_request;
-  expect_detach_child(*m_mock_imctx, mock_detach_child_request, 0);
-
-  MockMirrorDisableRequest mock_mirror_disable_request;
-  expect_mirror_disable(*m_mock_imctx, mock_mirror_disable_request, 0);
-
-  expect_state_close(*m_mock_imctx);
-
-  MockJournalRemoveRequest mock_journal_remove_request;
-  expect_journal_remove(*m_mock_imctx, mock_journal_remove_request, 0);
-
-  expect_remove_mirror_image(m_ioctx, 0);
-  expect_dir_remove_image(m_ioctx, 0);
-
-  C_SaferCond ctx;
-  librbd::NoOpProgressContext no_op;
-  ContextWQ op_work_queue;
-  MockRemoveRequest *req = MockRemoveRequest::create(
-    m_ioctx, m_image_name, "", true, false, no_op, &op_work_queue, &ctx);
-  req->send();
-
-  ASSERT_EQ(0, ctx.wait());
-
 }
 
 } // namespace image
